% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy_quarter.R
\name{fy_quarter}
\alias{fy_quarter}
\title{Return the financial quarter for the given month}
\usage{
fy_quarter(month)
}
\arguments{
\item{month}{}
}
\value{
financial quarter
}
\description{
Return the financial quarter for the given month
}
\examples{
data \%<>\% mutate(quarter = fy_quarter(lubridate::month(date)))
}
