% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round}
\alias{round}
\title{Function to mask the base round with janitor's round half up}
\usage{
round(x, digits = 0)
}
\arguments{
\item{x}{number to round}

\item{digits}{how many digits should be displayed after the decimal point?}
}
\value{
x rounded
}
\description{
Function to mask the base round with janitor's round half up
}
\examples{
round(0.5)
round(1.135, 2)
}
