% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_chi.R
\name{valid_chi}
\alias{valid_chi}
\title{Check to see if a CHI number is valid (not foolproof)}
\usage{
valid_chi(chi_number)
}
\arguments{
\item{chi_number}{}
}
\value{
boolean
}
\description{
Check to see if a CHI number is valid (not foolproof)
}
\examples{
data \%<>\% filter(valid_chi(chi))
}
