% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_postcode.R
\name{format_postcode}
\alias{format_postcode}
\title{Format postcodes into a consistent format}
\usage{
format_postcode(postcode_var, length = 7)
}
\arguments{
\item{length}{(Optional, default: 7-char)}
}
\value{
postcode
}
\description{
Format postcodes into a consistent format
}
\examples{
data \%<>\% mutate(postcode = format_postcode(postcode))
}
